Add files via upload
In this modified code, the IP address of the user is obtained using the $_SERVER['REMOTE_ADDR'] variable, and the current timestamp is obtained using the time() function. These two values are combined to create a unique identifier that is added to the file name. Finally, the file extension is extracted using the pathinfo() function and added back to the file name with the unique identifier.